+++ CSC8503/GameTech/TutorialGame.h
@@ -99,10 +99,10 @@ namespace NCL {
 				lockedObject = o;
 			}
 			GameUI* gameUI;
-			static void _openFirHandle(const EVENT* pEvent, UINT dwOwnerData);
-			static void _deleteHandle(const EVENT* pEvent, UINT dwOwnerData);
-			static void _HitHandle(const EVENT* pEvent, UINT dwOwnerData);
-			static void _respawnHandle(const EVENT* pEvent, UINT dwOwnerData);
+			static void _openFirHandle(const EVENT* pEvent, DWORD64 dwOwnerData);
+			static void _deleteHandle(const EVENT* pEvent, DWORD64 dwOwnerData);
+			static void _HitHandle(const EVENT* pEvent, DWORD64 dwOwnerData);
+			static void _respawnHandle(const EVENT* pEvent, DWORD64 dwOwnerData);
 			static TutorialGame* p_self;
 			static TutorialGame* getMe() { return p_self; }
 			void UpdateGameObjects(float dt);
